requirements.txt
python-telegram-bot
peewee
mysql-connector-python
bs4